Feat: 오븐 생성, 조회 기능 추가

+++ src/main/java/com/cookiepaper/dto/OvenDto.java
@@ -10,7 +10,7 @@ import lombok.*;
 public class OvenDto {
 
     @NotNull
-    private int ovId;
+    private Long ovId;
 
     @NotNull
     private String usId;
